@@ -6,29 +6,28 @@ publish_to: none
 
 dependencies:
   args: ^2.1.0
-  async: ^2.6.1
-  collection: ^1.15.0
+  async: ^2.10.0
+  collection: ^1.17.0
   colorize: ^3.0.0
-  file: ^6.1.0
-  # Pin git to 2.0.x until dart >=2.18 is legacy
-  git: '>=2.0.0 <2.1.0'
-  http: '>=0.13.3 <2.0.0'
+  file: ^7.0.1
+  git: ^2.0.0
+  http: ^1.0.0
   http_multi_server: ^3.0.1
-  meta: ^1.3.0
-  path: ^1.8.0
-  platform: ^3.0.0
+  meta: ^1.10.0
+  path: ^1.8.3
+  platform: ^3.0.2
   pub_semver: ^2.0.0
-  pubspec_parse: ^1.0.0
+  pubspec_parse: ^1.2.2
   quiver: ^3.0.1
-  test: ^1.17.3
-  uuid: ^3.0.4
+  test: ^1.24.0
+  uuid: ^4.5.1
   yaml: ^3.1.0
   yaml_edit: ^2.0.2
 
 dev_dependencies:
-  build_runner: ^2.0.3
-  matcher: ^0.12.10
+  build_runner: ^2.2.1
+  matcher: ^0.12.15
   mockito: ^5.4.4
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ^3.1.0
