@@ -218,12 +218,9 @@ void main() {
 
         // Emit test events
         final CameraClosingEvent event = CameraClosingEvent(cameraId);
-        await plugin.handleCameraMethodCall(
-            MethodCall('camera_closing', event.toJson()), cameraId);
-        await plugin.handleCameraMethodCall(
-            MethodCall('camera_closing', event.toJson()), cameraId);
-        await plugin.handleCameraMethodCall(
-            MethodCall('camera_closing', event.toJson()), cameraId);
+        plugin.hostCameraHandlers[cameraId]!.cameraClosing();
+        plugin.hostCameraHandlers[cameraId]!.cameraClosing();
+        plugin.hostCameraHandlers[cameraId]!.cameraClosing();
 
         // Assert
         expect(await streamQueue.next, event);
@@ -242,14 +239,11 @@ void main() {
             StreamQueue<CameraErrorEvent>(errorStream);
 
         // Emit test events
-        final CameraErrorEvent event =
-            CameraErrorEvent(cameraId, 'Error Description');
-        await plugin.handleCameraMethodCall(
-            MethodCall('error', event.toJson()), cameraId);
-        await plugin.handleCameraMethodCall(
-            MethodCall('error', event.toJson()), cameraId);
-        await plugin.handleCameraMethodCall(
-            MethodCall('error', event.toJson()), cameraId);
+        const String errorMessage = 'Error Description';
+        final CameraErrorEvent event = CameraErrorEvent(cameraId, errorMessage);
+        plugin.hostCameraHandlers[cameraId]!.error(errorMessage);
+        plugin.hostCameraHandlers[cameraId]!.error(errorMessage);
+        plugin.hostCameraHandlers[cameraId]!.error(errorMessage);
 
         // Assert
         expect(await streamQueue.next, event);
@@ -489,15 +483,6 @@ void main() {
         // Act
         expect(widget is Texture, isTrue);
         expect((widget as Texture).textureId, cameraId);
-      });
-
-      test('Should throw UnimplementedError when handling unknown method', () {
-        final CameraWindows plugin = CameraWindows(api: mockApi);
-
-        expect(
-            () => plugin.handleCameraMethodCall(
-                const MethodCall('unknown_method'), 1),
-            throwsA(isA<UnimplementedError>()));
       });
 
       test('Should get the max zoom level', () async {
