@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v21.0.0), do not edit directly.
+// Autogenerated from Pigeon (v22.6.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 #undef _HAS_EXCEPTIONS
@@ -127,15 +127,16 @@ PlatformMediaSettings PlatformMediaSettings::FromEncodableList(
                                 std::get<bool>(list[4]));
   auto& encodable_frames_per_second = list[1];
   if (!encodable_frames_per_second.IsNull()) {
-    decoded.set_frames_per_second(encodable_frames_per_second.LongValue());
+    decoded.set_frames_per_second(
+        std::get<int64_t>(encodable_frames_per_second));
   }
   auto& encodable_video_bitrate = list[2];
   if (!encodable_video_bitrate.IsNull()) {
-    decoded.set_video_bitrate(encodable_video_bitrate.LongValue());
+    decoded.set_video_bitrate(std::get<int64_t>(encodable_video_bitrate));
   }
   auto& encodable_audio_bitrate = list[3];
   if (!encodable_audio_bitrate.IsNull()) {
-    decoded.set_audio_bitrate(encodable_audio_bitrate.LongValue());
+    decoded.set_audio_bitrate(std::get<int64_t>(encodable_audio_bitrate));
   }
   return decoded;
 }
@@ -166,18 +167,12 @@ PlatformSize PlatformSize::FromEncodableList(const EncodableList& list) {
   return decoded;
 }
 
-PigeonCodecSerializer::PigeonCodecSerializer() {}
+PigeonInternalCodecSerializer::PigeonInternalCodecSerializer() {}
 
-EncodableValue PigeonCodecSerializer::ReadValueOfType(
+EncodableValue PigeonInternalCodecSerializer::ReadValueOfType(
     uint8_t type, flutter::ByteStreamReader* stream) const {
   switch (type) {
-    case 129:
-      return CustomEncodableValue(PlatformMediaSettings::FromEncodableList(
-          std::get<EncodableList>(ReadValue(stream))));
-    case 130:
-      return CustomEncodableValue(PlatformSize::FromEncodableList(
-          std::get<EncodableList>(ReadValue(stream))));
-    case 131: {
+    case 129: {
       const auto& encodable_enum_arg = ReadValue(stream);
       const int64_t enum_arg_value =
           encodable_enum_arg.IsNull() ? 0 : encodable_enum_arg.LongValue();
@@ -186,17 +181,32 @@ EncodableValue PigeonCodecSerializer::ReadValueOfType(
                  : CustomEncodableValue(
                        static_cast<PlatformResolutionPreset>(enum_arg_value));
     }
+    case 130: {
+      return CustomEncodableValue(PlatformMediaSettings::FromEncodableList(
+          std::get<EncodableList>(ReadValue(stream))));
+    }
+    case 131: {
+      return CustomEncodableValue(PlatformSize::FromEncodableList(
+          std::get<EncodableList>(ReadValue(stream))));
+    }
     default:
       return flutter::StandardCodecSerializer::ReadValueOfType(type, stream);
   }
 }
 
-void PigeonCodecSerializer::WriteValue(
+void PigeonInternalCodecSerializer::WriteValue(
     const EncodableValue& value, flutter::ByteStreamWriter* stream) const {
   if (const CustomEncodableValue* custom_value =
           std::get_if<CustomEncodableValue>(&value)) {
-    if (custom_value->type() == typeid(PlatformMediaSettings)) {
+    if (custom_value->type() == typeid(PlatformResolutionPreset)) {
       stream->WriteByte(129);
+      WriteValue(EncodableValue(static_cast<int>(
+                     std::any_cast<PlatformResolutionPreset>(*custom_value))),
+                 stream);
+      return;
+    }
+    if (custom_value->type() == typeid(PlatformMediaSettings)) {
+      stream->WriteByte(130);
       WriteValue(
           EncodableValue(std::any_cast<PlatformMediaSettings>(*custom_value)
                              .ToEncodableList()),
@@ -204,18 +214,11 @@ void PigeonCodecSerializer::WriteValue(
       return;
     }
     if (custom_value->type() == typeid(PlatformSize)) {
-      stream->WriteByte(130);
+      stream->WriteByte(131);
       WriteValue(
           EncodableValue(
               std::any_cast<PlatformSize>(*custom_value).ToEncodableList()),
           stream);
-      return;
-    }
-    if (custom_value->type() == typeid(PlatformResolutionPreset)) {
-      stream->WriteByte(131);
-      WriteValue(EncodableValue(static_cast<int>(
-                     std::any_cast<PlatformResolutionPreset>(*custom_value))),
-                 stream);
       return;
     }
   }
@@ -225,7 +228,7 @@ void PigeonCodecSerializer::WriteValue(
 /// The codec used by CameraApi.
 const flutter::StandardMessageCodec& CameraApi::GetCodec() {
   return flutter::StandardMessageCodec::GetInstance(
-      &PigeonCodecSerializer::GetInstance());
+      &PigeonInternalCodecSerializer::GetInstance());
 }
 
 // Sets up an instance of `CameraApi` to handle messages through the
@@ -498,78 +501,6 @@ void CameraApi::SetUp(flutter::BinaryMessenger* binary_messenger,
   {
     BasicMessageChannel<> channel(
         binary_messenger,
-        "dev.flutter.pigeon.camera_windows.CameraApi.startImageStream" +
-            prepended_suffix,
-        &GetCodec());
-    if (api != nullptr) {
-      channel.SetMessageHandler(
-          [api](const EncodableValue& message,
-                const flutter::MessageReply<EncodableValue>& reply) {
-            try {
-              const auto& args = std::get<EncodableList>(message);
-              const auto& encodable_camera_id_arg = args.at(0);
-              if (encodable_camera_id_arg.IsNull()) {
-                reply(WrapError("camera_id_arg unexpectedly null."));
-                return;
-              }
-              const int64_t camera_id_arg = encodable_camera_id_arg.LongValue();
-              api->StartImageStream(
-                  camera_id_arg, [reply](std::optional<FlutterError>&& output) {
-                    if (output.has_value()) {
-                      reply(WrapError(output.value()));
-                      return;
-                    }
-                    EncodableList wrapped;
-                    wrapped.push_back(EncodableValue());
-                    reply(EncodableValue(std::move(wrapped)));
-                  });
-            } catch (const std::exception& exception) {
-              reply(WrapError(exception.what()));
-            }
-          });
-    } else {
-      channel.SetMessageHandler(nullptr);
-    }
-  }
-  {
-    BasicMessageChannel<> channel(
-        binary_messenger,
-        "dev.flutter.pigeon.camera_windows.CameraApi.stopImageStream" +
-            prepended_suffix,
-        &GetCodec());
-    if (api != nullptr) {
-      channel.SetMessageHandler(
-          [api](const EncodableValue& message,
-                const flutter::MessageReply<EncodableValue>& reply) {
-            try {
-              const auto& args = std::get<EncodableList>(message);
-              const auto& encodable_camera_id_arg = args.at(0);
-              if (encodable_camera_id_arg.IsNull()) {
-                reply(WrapError("camera_id_arg unexpectedly null."));
-                return;
-              }
-              const int64_t camera_id_arg = encodable_camera_id_arg.LongValue();
-              api->StopImageStream(
-                  camera_id_arg, [reply](std::optional<FlutterError>&& output) {
-                    if (output.has_value()) {
-                      reply(WrapError(output.value()));
-                      return;
-                    }
-                    EncodableList wrapped;
-                    wrapped.push_back(EncodableValue());
-                    reply(EncodableValue(std::move(wrapped)));
-                  });
-            } catch (const std::exception& exception) {
-              reply(WrapError(exception.what()));
-            }
-          });
-    } else {
-      channel.SetMessageHandler(nullptr);
-    }
-  }
-  {
-    BasicMessageChannel<> channel(
-        binary_messenger,
         "dev.flutter.pigeon.camera_windows.CameraApi.pausePreview" +
             prepended_suffix,
         &GetCodec());
@@ -651,6 +582,90 @@ EncodableValue CameraApi::WrapError(const FlutterError& error) {
   return EncodableValue(EncodableList{EncodableValue(error.code()),
                                       EncodableValue(error.message()),
                                       error.details()});
+}
+
+// Generated class from Pigeon that represents Flutter messages that can be
+// called from C++.
+CameraEventApi::CameraEventApi(flutter::BinaryMessenger* binary_messenger)
+    : binary_messenger_(binary_messenger), message_channel_suffix_("") {}
+
+CameraEventApi::CameraEventApi(flutter::BinaryMessenger* binary_messenger,
+                               const std::string& message_channel_suffix)
+    : binary_messenger_(binary_messenger),
+      message_channel_suffix_(message_channel_suffix.length() > 0
+                                  ? std::string(".") + message_channel_suffix
+                                  : "") {}
+
+const flutter::StandardMessageCodec& CameraEventApi::GetCodec() {
+  return flutter::StandardMessageCodec::GetInstance(
+      &PigeonInternalCodecSerializer::GetInstance());
+}
+
+void CameraEventApi::CameraClosing(
+    std::function<void(void)>&& on_success,
+    std::function<void(const FlutterError&)>&& on_error) {
+  const std::string channel_name =
+      "dev.flutter.pigeon.camera_windows.CameraEventApi.cameraClosing" +
+      message_channel_suffix_;
+  BasicMessageChannel<> channel(binary_messenger_, channel_name, &GetCodec());
+  EncodableValue encoded_api_arguments = EncodableValue();
+  channel.Send(
+      encoded_api_arguments, [channel_name, on_success = std::move(on_success),
+                              on_error = std::move(on_error)](
+                                 const uint8_t* reply, size_t reply_size) {
+        std::unique_ptr<EncodableValue> response =
+            GetCodec().DecodeMessage(reply, reply_size);
+        const auto& encodable_return_value = *response;
+        const auto* list_return_value =
+            std::get_if<EncodableList>(&encodable_return_value);
+        if (list_return_value) {
+          if (list_return_value->size() > 1) {
+            on_error(
+                FlutterError(std::get<std::string>(list_return_value->at(0)),
+                             std::get<std::string>(list_return_value->at(1)),
+                             list_return_value->at(2)));
+          } else {
+            on_success();
+          }
+        } else {
+          on_error(CreateConnectionError(channel_name));
+        }
+      });
+}
+
+void CameraEventApi::Error(
+    const std::string& error_message_arg,
+    std::function<void(void)>&& on_success,
+    std::function<void(const FlutterError&)>&& on_error) {
+  const std::string channel_name =
+      "dev.flutter.pigeon.camera_windows.CameraEventApi.error" +
+      message_channel_suffix_;
+  BasicMessageChannel<> channel(binary_messenger_, channel_name, &GetCodec());
+  EncodableValue encoded_api_arguments = EncodableValue(EncodableList{
+      EncodableValue(error_message_arg),
+  });
+  channel.Send(
+      encoded_api_arguments, [channel_name, on_success = std::move(on_success),
+                              on_error = std::move(on_error)](
+                                 const uint8_t* reply, size_t reply_size) {
+        std::unique_ptr<EncodableValue> response =
+            GetCodec().DecodeMessage(reply, reply_size);
+        const auto& encodable_return_value = *response;
+        const auto* list_return_value =
+            std::get_if<EncodableList>(&encodable_return_value);
+        if (list_return_value) {
+          if (list_return_value->size() > 1) {
+            on_error(
+                FlutterError(std::get<std::string>(list_return_value->at(0)),
+                             std::get<std::string>(list_return_value->at(1)),
+                             list_return_value->at(2)));
+          } else {
+            on_success();
+          }
+        } else {
+          on_error(CreateConnectionError(channel_name));
+        }
+      });
 }
 
 }  // namespace camera_windows
