@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v21.0.0), do not edit directly.
+// Autogenerated from Pigeon (v22.6.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 #ifndef PIGEON_MESSAGES_G_H_
@@ -52,6 +52,7 @@ class ErrorOr {
 
  private:
   friend class CameraApi;
+  friend class CameraEventApi;
   ErrorOr() = default;
   T TakeValue() && { return std::get<T>(std::move(v_)); }
 
@@ -106,7 +107,8 @@ class PlatformMediaSettings {
       const flutter::EncodableList& list);
   flutter::EncodableList ToEncodableList() const;
   friend class CameraApi;
-  friend class PigeonCodecSerializer;
+  friend class CameraEventApi;
+  friend class PigeonInternalCodecSerializer;
   PlatformResolutionPreset resolution_preset_;
   std::optional<int64_t> frames_per_second_;
   std::optional<int64_t> video_bitrate_;
@@ -132,16 +134,17 @@ class PlatformSize {
   static PlatformSize FromEncodableList(const flutter::EncodableList& list);
   flutter::EncodableList ToEncodableList() const;
   friend class CameraApi;
-  friend class PigeonCodecSerializer;
+  friend class CameraEventApi;
+  friend class PigeonInternalCodecSerializer;
   double width_;
   double height_;
 };
 
-class PigeonCodecSerializer : public flutter::StandardCodecSerializer {
+class PigeonInternalCodecSerializer : public flutter::StandardCodecSerializer {
  public:
-  PigeonCodecSerializer();
-  inline static PigeonCodecSerializer& GetInstance() {
-    static PigeonCodecSerializer sInstance;
+  PigeonInternalCodecSerializer();
+  inline static PigeonInternalCodecSerializer& GetInstance() {
+    static PigeonInternalCodecSerializer sInstance;
     return sInstance;
   }
 
@@ -186,14 +189,6 @@ class CameraApi {
   virtual void StopVideoRecording(
       int64_t camera_id,
       std::function<void(ErrorOr<std::string> reply)> result) = 0;
-  // Starts the image stream for the given camera.
-  virtual void StartImageStream(
-      int64_t camera_id,
-      std::function<void(std::optional<FlutterError> reply)> result) = 0;
-  // Stops the image stream for the given camera.
-  virtual void StopImageStream(
-      int64_t camera_id,
-      std::function<void(std::optional<FlutterError> reply)> result) = 0;
   // Starts the preview stream for the given camera.
   virtual void PausePreview(
       int64_t camera_id,
@@ -216,5 +211,26 @@ class CameraApi {
  protected:
   CameraApi() = default;
 };
+// Generated class from Pigeon that represents Flutter messages that can be
+// called from C++.
+class CameraEventApi {
+ public:
+  CameraEventApi(flutter::BinaryMessenger* binary_messenger);
+  CameraEventApi(flutter::BinaryMessenger* binary_messenger,
+                 const std::string& message_channel_suffix);
+  static const flutter::StandardMessageCodec& GetCodec();
+  // Called when the camera instance is closing on the native side.
+  void CameraClosing(std::function<void(void)>&& on_success,
+                     std::function<void(const FlutterError&)>&& on_error);
+  // Called when a camera error occurs on the native side.
+  void Error(const std::string& error_message,
+             std::function<void(void)>&& on_success,
+             std::function<void(const FlutterError&)>&& on_error);
+
+ private:
+  flutter::BinaryMessenger* binary_messenger_;
+  std::string message_channel_suffix_;
+};
+
 }  // namespace camera_windows
 #endif  // PIGEON_MESSAGES_G_H_
