@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v22.4.2), do not edit directly.
+// Autogenerated from Pigeon (v22.6.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 import Foundation
@@ -315,6 +315,7 @@ struct SK2TransactionMessage {
   var purchasedQuantity: Int64
   var appAccountToken: String? = nil
   var restoring: Bool
+  var receiptData: String? = nil
   var error: SK2ErrorMessage? = nil
 
   // swift-format-ignore: AlwaysUseLowerCamelCase
@@ -326,7 +327,8 @@ struct SK2TransactionMessage {
     let purchasedQuantity = pigeonVar_list[4] as! Int64
     let appAccountToken: String? = nilOrValue(pigeonVar_list[5])
     let restoring = pigeonVar_list[6] as! Bool
-    let error: SK2ErrorMessage? = nilOrValue(pigeonVar_list[7])
+    let receiptData: String? = nilOrValue(pigeonVar_list[7])
+    let error: SK2ErrorMessage? = nilOrValue(pigeonVar_list[8])
 
     return SK2TransactionMessage(
       id: id,
@@ -336,6 +338,7 @@ struct SK2TransactionMessage {
       purchasedQuantity: purchasedQuantity,
       appAccountToken: appAccountToken,
       restoring: restoring,
+      receiptData: receiptData,
       error: error
     )
   }
@@ -348,6 +351,7 @@ struct SK2TransactionMessage {
       purchasedQuantity,
       appAccountToken,
       restoring,
+      receiptData,
       error,
     ]
   }
@@ -508,6 +512,7 @@ protocol InAppPurchase2API {
   func finish(id: Int64, completion: @escaping (Result<Void, Error>) -> Void)
   func startListeningToTransactions() throws
   func stopListeningToTransactions() throws
+  func restorePurchases(completion: @escaping (Result<Void, Error>) -> Void)
 }
 
 /// Generated setup class from Pigeon to handle messages through the `binaryMessenger`.
@@ -645,12 +650,30 @@ class InAppPurchase2APISetup {
     } else {
       stopListeningToTransactionsChannel.setMessageHandler(nil)
     }
+    let restorePurchasesChannel = FlutterBasicMessageChannel(
+      name:
+        "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.restorePurchases\(channelSuffix)",
+      binaryMessenger: binaryMessenger, codec: codec)
+    if let api = api {
+      restorePurchasesChannel.setMessageHandler { _, reply in
+        api.restorePurchases { result in
+          switch result {
+          case .success:
+            reply(wrapResult(nil))
+          case .failure(let error):
+            reply(wrapError(error))
+          }
+        }
+      }
+    } else {
+      restorePurchasesChannel.setMessageHandler(nil)
+    }
   }
 }
 /// Generated protocol from Pigeon that represents Flutter messages that can be called from Swift.
 protocol InAppPurchase2CallbackAPIProtocol {
   func onTransactionsUpdated(
-    newTransaction newTransactionArg: SK2TransactionMessage,
+    newTransactions newTransactionsArg: [SK2TransactionMessage],
     completion: @escaping (Result<Void, PigeonError>) -> Void)
 }
 class InAppPurchase2CallbackAPI: InAppPurchase2CallbackAPIProtocol {
@@ -664,14 +687,14 @@ class InAppPurchase2CallbackAPI: InAppPurchase2CallbackAPIProtocol {
     return sk2_pigeonPigeonCodec.shared
   }
   func onTransactionsUpdated(
-    newTransaction newTransactionArg: SK2TransactionMessage,
+    newTransactions newTransactionsArg: [SK2TransactionMessage],
     completion: @escaping (Result<Void, PigeonError>) -> Void
   ) {
     let channelName: String =
       "dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2CallbackAPI.onTransactionsUpdated\(messageChannelSuffix)"
     let channel = FlutterBasicMessageChannel(
       name: channelName, binaryMessenger: binaryMessenger, codec: codec)
-    channel.sendMessage([newTransactionArg] as [Any?]) { response in
+    channel.sendMessage([newTransactionsArg] as [Any?]) { response in
       guard let listResponse = response as? [Any?] else {
         completion(.failure(createConnectionError(withChannelName: channelName)))
         return
