@@ -1,7 +1,7 @@
 // Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-// Autogenerated from Pigeon (v22.4.2), do not edit directly.
+// Autogenerated from Pigeon (v22.6.0), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 // ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import, no_leading_underscores_for_local_identifiers
 
@@ -303,6 +303,7 @@ class SK2TransactionMessage {
     this.purchasedQuantity = 1,
     this.appAccountToken,
     this.restoring = false,
+    this.receiptData,
     this.error,
   });
 
@@ -320,6 +321,8 @@ class SK2TransactionMessage {
 
   bool restoring;
 
+  String? receiptData;
+
   SK2ErrorMessage? error;
 
   Object encode() {
@@ -331,6 +334,7 @@ class SK2TransactionMessage {
       purchasedQuantity,
       appAccountToken,
       restoring,
+      receiptData,
       error,
     ];
   }
@@ -345,7 +349,8 @@ class SK2TransactionMessage {
       purchasedQuantity: result[4]! as int,
       appAccountToken: result[5] as String?,
       restoring: result[6]! as bool,
-      error: result[7] as SK2ErrorMessage?,
+      receiptData: result[7] as String?,
+      error: result[8] as SK2ErrorMessage?,
     );
   }
 }
@@ -685,12 +690,36 @@ class InAppPurchase2API {
       return;
     }
   }
+
+  Future<void> restorePurchases() async {
+    final String pigeonVar_channelName =
+        'dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2API.restorePurchases$pigeonVar_messageChannelSuffix';
+    final BasicMessageChannel<Object?> pigeonVar_channel =
+        BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
+      pigeonChannelCodec,
+      binaryMessenger: pigeonVar_binaryMessenger,
+    );
+    final List<Object?>? pigeonVar_replyList =
+        await pigeonVar_channel.send(null) as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
+      throw PlatformException(
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
+      );
+    } else {
+      return;
+    }
+  }
 }
 
 abstract class InAppPurchase2CallbackAPI {
   static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
 
-  void onTransactionsUpdated(SK2TransactionMessage newTransaction);
+  void onTransactionsUpdated(List<SK2TransactionMessage> newTransactions);
 
   static void setUp(
     InAppPurchase2CallbackAPI? api, {
@@ -713,12 +742,12 @@ abstract class InAppPurchase2CallbackAPI {
           assert(message != null,
               'Argument for dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2CallbackAPI.onTransactionsUpdated was null.');
           final List<Object?> args = (message as List<Object?>?)!;
-          final SK2TransactionMessage? arg_newTransaction =
-              (args[0] as SK2TransactionMessage?);
-          assert(arg_newTransaction != null,
-              'Argument for dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2CallbackAPI.onTransactionsUpdated was null, expected non-null SK2TransactionMessage.');
+          final List<SK2TransactionMessage>? arg_newTransactions =
+              (args[0] as List<Object?>?)?.cast<SK2TransactionMessage>();
+          assert(arg_newTransactions != null,
+              'Argument for dev.flutter.pigeon.in_app_purchase_storekit.InAppPurchase2CallbackAPI.onTransactionsUpdated was null, expected non-null List<SK2TransactionMessage>.');
           try {
-            api.onTransactionsUpdated(arg_newTransaction!);
+            api.onTransactionsUpdated(arg_newTransactions!);
             return wrapResponse(empty: true);
           } on PlatformException catch (e) {
             return wrapResponse(error: e);
