@@ -144,6 +144,7 @@ class SK2TransactionMessage {
       this.purchasedQuantity = 1,
       this.appAccountToken,
       this.error,
+      this.receiptData,
       this.restoring = false});
   final int id;
   final int originalId;
@@ -152,6 +153,7 @@ class SK2TransactionMessage {
   final int purchasedQuantity;
   final String? appAccountToken;
   final bool restoring;
+  final String? receiptData;
   final SK2ErrorMessage? error;
 }
 
@@ -189,9 +191,12 @@ abstract class InAppPurchase2API {
   void startListeningToTransactions();
 
   void stopListeningToTransactions();
+
+  @async
+  void restorePurchases();
 }
 
 @FlutterApi()
 abstract class InAppPurchase2CallbackAPI {
-  void onTransactionsUpdated(SK2TransactionMessage newTransaction);
+  void onTransactionsUpdated(List<SK2TransactionMessage> newTransactions);
 }
