@@ -6,7 +6,9 @@ import 'dart:convert';
 import 'dart:io' as io;
 
 import 'package:file/file.dart';
+import 'package:platform/platform.dart';
 
+import 'core.dart';
 import 'output_utils.dart';
 import 'process_runner.dart';
 
@@ -33,36 +35,81 @@ class Xcode {
   /// Runs an `xcodebuild` in [directory] with the given parameters.
   Future<int> runXcodeBuild(
     Directory exampleDirectory,
-    String platform, {
+    String targetPlatform, {
     List<String> actions = const <String>['build'],
     required String workspace,
     required String scheme,
     String? configuration,
     List<String> extraFlags = const <String>[],
-  }) {
-    File? disabledSandboxEntitlementFile;
-    if (actions.contains('test') && platform.toLowerCase() == 'macos') {
-      disabledSandboxEntitlementFile = _createDisabledSandboxEntitlementFile(
-        exampleDirectory.childDirectory(platform.toLowerCase()),
-        configuration ?? 'Debug',
-      );
+    required Platform hostPlatform,
+  }) async {
+    final FileSystem fileSystem = exampleDirectory.fileSystem;
+    String? resultBundlePath;
+    final Directory? logsDirectory = ciLogsDirectory(hostPlatform, fileSystem);
+    Directory? resultBundleTemp;
+    try {
+      if (logsDirectory != null) {
+        resultBundleTemp =
+            fileSystem.systemTempDirectory.createTempSync('flutter_xcresult.');
+        resultBundlePath = resultBundleTemp.childDirectory('result').path;
+      }
+      File? disabledSandboxEntitlementFile;
+      if (actions.contains('test') && targetPlatform.toLowerCase() == 'macos') {
+        disabledSandboxEntitlementFile = _createDisabledSandboxEntitlementFile(
+          exampleDirectory.childDirectory(targetPlatform.toLowerCase()),
+          configuration ?? 'Debug',
+        );
+      }
+      final List<String> args = <String>[
+        _xcodeBuildCommand,
+        ...actions,
+        ...<String>['-workspace', workspace],
+        ...<String>['-scheme', scheme],
+        if (resultBundlePath != null) ...<String>[
+          '-resultBundlePath',
+          resultBundlePath
+        ],
+        if (configuration != null) ...<String>['-configuration', configuration],
+        ...extraFlags,
+        if (disabledSandboxEntitlementFile != null)
+          'CODE_SIGN_ENTITLEMENTS=${disabledSandboxEntitlementFile.path}',
+      ];
+      final String completeTestCommand = '$_xcRunCommand ${args.join(' ')}';
+      if (log) {
+        print(completeTestCommand);
+      }
+      final int resultExit = await processRunner
+          .runAndStream(_xcRunCommand, args, workingDir: exampleDirectory);
+
+      if (resultExit != 0 && resultBundleTemp != null) {
+        final Directory xcresultBundle =
+            resultBundleTemp.childDirectory('result.xcresult');
+        if (logsDirectory != null) {
+          if (xcresultBundle.existsSync()) {
+            // Zip the test results to the artifacts directory for upload.
+            final File zipPath = logsDirectory.childFile(
+                'xcodebuild-${DateTime.now().toLocal().toIso8601String()}.zip');
+            await processRunner.run(
+              'zip',
+              <String>[
+                '-r',
+                '-9',
+                '-q',
+                zipPath.path,
+                xcresultBundle.basename,
+              ],
+              workingDir: resultBundleTemp,
+            );
+          } else {
+            print(
+                'xcresult bundle ${xcresultBundle.path} does not exist, skipping upload');
+          }
+        }
+      }
+      return resultExit;
+    } finally {
+      resultBundleTemp?.deleteSync(recursive: true);
     }
-    final List<String> args = <String>[
-      _xcodeBuildCommand,
-      ...actions,
-      ...<String>['-workspace', workspace],
-      ...<String>['-scheme', scheme],
-      if (configuration != null) ...<String>['-configuration', configuration],
-      ...extraFlags,
-      if (disabledSandboxEntitlementFile != null)
-        'CODE_SIGN_ENTITLEMENTS=${disabledSandboxEntitlementFile.path}',
-    ];
-    final String completeTestCommand = '$_xcRunCommand ${args.join(' ')}';
-    if (log) {
-      print(completeTestCommand);
-    }
-    return processRunner.runAndStream(_xcRunCommand, args,
-        workingDir: exampleDirectory);
   }
 
   /// Returns true if [project], which should be an .xcodeproj directory,
