@@ -202,7 +202,6 @@ class GradleCheckCommand extends PackageLoopingCommand {
   /// configuration that enables artifact hub env variable.
   @visibleForTesting
   static String exampleRootSettingsArtifactHubString = '''
-// See $artifactHubDocumentationString for more info.
 buildscript {
   repositories {
     maven {
@@ -216,6 +215,18 @@ buildscript {
 apply plugin: "com.google.cloud.artifactregistry.gradle-plugin"
 ''';
 
+  /// String printed as a valid example of settings.gradle repository
+  /// configuration that enables artifact hub env variable.
+  /// GP stands for the gradle plugin method of flutter tooling inclusion.
+  @visibleForTesting
+  static String exampleSettingsArtifactHubStringGP = '''
+plugins {
+    id "dev.flutter.flutter-plugin-loader" version "1.0.0"
+    // ...other plugins
+    id "com.google.cloud.artifactregistry.gradle-plugin" version "2.2.1"
+}
+  ''';
+
   /// Validates that [gradleLines] reads and uses a artifiact hub repository
   /// when ARTIFACT_HUB_REPOSITORY is set.
   ///
@@ -228,6 +239,10 @@ apply plugin: "com.google.cloud.artifactregistry.gradle-plugin"
         r'classpath.*gradle\.plugin\.com\.google\.cloud\.artifactregistry:artifactregistry-gradle-plugin');
     final RegExp artifactRegistryPluginApplyRegex = RegExp(
         r'apply.*plugin.*com\.google\.cloud\.artifactregistry\.gradle-plugin');
+    final RegExp artifactRegistryPluginApplyRegexGP = RegExp(
+        r'id.*com\.google\.cloud\.artifactregistry\.gradle-plugin.*version.*\b\d+\.\d+\.\d+\b');
+    final RegExp artifactRegistryPluginApplyDeclarativeRegex =
+        RegExp(r'\bpluginManagement\b');
 
     final bool documentationPresent = gradleLines
         .any((String line) => documentationPresentRegex.hasMatch(line));
@@ -235,17 +250,36 @@ apply plugin: "com.google.cloud.artifactregistry.gradle-plugin"
         .any((String line) => artifactRegistryDefinitionRegex.hasMatch(line));
     final bool artifactRegistryPluginApplied = gradleLines
         .any((String line) => artifactRegistryPluginApplyRegex.hasMatch(line));
+    final bool declarativeArtifactRegistryApplied = gradleLines.any(
+        (String line) => artifactRegistryPluginApplyRegexGP.hasMatch(line));
+    final bool declarativePluginBlockApplied = gradleLines.any((String line) =>
+        artifactRegistryPluginApplyDeclarativeRegex.hasMatch(line));
 
-    if (!(documentationPresent &&
-        artifactRegistryDefined &&
-        artifactRegistryPluginApplied)) {
-      printError('Failed Artifact Hub validation. Include the following in '
-          'example root settings.gradle:\n$exampleRootSettingsArtifactHubString');
+    final bool imperativeArtifactRegistryApplied =
+        artifactRegistryDefined && artifactRegistryPluginApplied;
+
+    final bool validArtifactConfiguration = documentationPresent &&
+        (imperativeArtifactRegistryApplied ||
+            declarativeArtifactRegistryApplied);
+
+    if (!validArtifactConfiguration) {
+      printError('Failed Artifact Hub validation.');
+      if (!documentationPresent) {
+        printError(
+            'The link to the Artifact Hub documentation is missing. Include the following in '
+            'example root settings.gradle:\n// See $artifactHubDocumentationString for more info.');
+      }
+      if (artifactRegistryDefined ||
+          artifactRegistryPluginApplied ||
+          !declarativePluginBlockApplied) {
+        printError('Include the following in '
+            'example root settings.gradle:\n$exampleRootSettingsArtifactHubString');
+      } else if (!declarativeArtifactRegistryApplied) {
+        printError('Include the following in '
+            'example root settings.gradle:\n$exampleSettingsArtifactHubStringGP');
+      }
     }
-
-    return documentationPresent &&
-        artifactRegistryDefined &&
-        artifactRegistryPluginApplied;
+    return validArtifactConfiguration;
   }
 
   /// Validates the top-level build.gradle for an example app (e.g.,
