@@ -174,6 +174,7 @@ ${warningsConfigured ? warningConfig : ''}
   void writeFakeExampleTopLevelSettingsGradle(
     RepositoryPackage package, {
     bool includeArtifactHub = true,
+    bool includeArtifactDocumentation = true,
   }) {
     final File settingsGradle = package
         .platformDirectory(FlutterPlatform.android)
@@ -196,7 +197,55 @@ plugins.each { name, path ->
     include ":\$name"
     project(":\$name").projectDir = pluginDirectory
 }
+${includeArtifactDocumentation ? '// See ${GradleCheckCommand.artifactHubDocumentationString} for more info.' : ''}
 ${includeArtifactHub ? GradleCheckCommand.exampleRootSettingsArtifactHubString : ''}
+''');
+  }
+
+  /// Writes a fake android/build.gradle file for an example [package] with the
+  /// given options.
+  void writeFakeExampleSettingsGradle(
+    RepositoryPackage package, {
+    bool includeArtifactHub = true,
+    bool includeArtifactDocumentation = true,
+  }) {
+    final File settingsGradle = package
+        .platformDirectory(FlutterPlatform.android)
+        .childFile('settings.gradle');
+    settingsGradle.createSync(recursive: true);
+
+    /// String printed as a valid example of settings.gradle repository
+    /// configuration without the artifact hub env variable.
+    /// GP stands for the gradle plugin method of flutter tooling inclusion.
+    const String exampleSettingsWithoutArtifactHubStringGP = '''
+plugins {
+    id "dev.flutter.flutter-plugin-loader" version "1.0.0"
+    // ...other plugins
+}
+  ''';
+
+    settingsGradle.writeAsStringSync('''
+pluginManagement {
+  def flutterSdkPath = {
+    def properties = new Properties()
+    file("local.properties").withInputStream { properties.load(it) }
+    def flutterSdkPath = properties.getProperty("flutter.sdk")
+    assert flutterSdkPath != null, "flutter.sdk not set in local.properties"
+    return flutterSdkPath
+  }()
+
+  includeBuild("\$flutterSdkPath/packages/flutter_tools/gradle")
+
+  repositories {
+    google()
+    mavenCentral()
+    gradlePluginPortal()
+  }
+}
+
+${includeArtifactDocumentation ? '// See ${GradleCheckCommand.artifactHubDocumentationString} for more info.' : ''}
+${includeArtifactHub ? GradleCheckCommand.exampleSettingsArtifactHubStringGP : exampleSettingsWithoutArtifactHubStringGP}
+include ":app"
 ''');
   }
 
@@ -258,6 +307,7 @@ dependencies {
     String? kotlinVersion,
     bool includeBuildArtifactHub = true,
     bool includeSettingsArtifactHub = true,
+    bool includeSettingsDocumentationArtifactHub = true,
   }) {
     writeFakeExampleTopLevelBuildGradle(
       package,
@@ -271,6 +321,34 @@ dependencies {
     writeFakeExampleTopLevelSettingsGradle(
       package,
       includeArtifactHub: includeSettingsArtifactHub,
+      includeArtifactDocumentation: includeSettingsDocumentationArtifactHub,
+    );
+  }
+
+  void writeFakeExampleBuildGradleGP(
+    RepositoryPackage package, {
+    required String pluginName,
+    bool includeNamespace = true,
+    bool commentNamespace = false,
+    bool warningsConfigured = true,
+    String? kotlinVersion,
+    required bool includeBuildArtifactHub,
+    required bool includeSettingsArtifactHub,
+    required bool includeSettingsDocumentationArtifactHub,
+  }) {
+    writeFakeExampleTopLevelBuildGradle(
+      package,
+      pluginName: pluginName,
+      warningsConfigured: warningsConfigured,
+      kotlinVersion: kotlinVersion,
+      includeArtifactHub: includeBuildArtifactHub,
+    );
+    writeFakeExampleAppBuildGradle(package,
+        includeNamespace: includeNamespace, commentNamespace: commentNamespace);
+    writeFakeExampleSettingsGradle(
+      package,
+      includeArtifactHub: includeSettingsArtifactHub,
+      includeArtifactDocumentation: includeSettingsDocumentationArtifactHub,
     );
   }
 
@@ -670,14 +748,14 @@ dependencies {
       writeFakePluginBuildGradle(package, includeLanguageVersion: true);
       writeFakeManifest(package);
       final RepositoryPackage example = package.getExamples().first;
-      writeFakeExampleBuildGradles(
-        example,
-        pluginName: packageName,
-        // ignore: avoid_redundant_argument_values
-        includeBuildArtifactHub: true,
-        // ignore: avoid_redundant_argument_values
-        includeSettingsArtifactHub: true,
-      );
+      writeFakeExampleBuildGradles(example,
+          pluginName: packageName,
+          // ignore: avoid_redundant_argument_values
+          includeBuildArtifactHub: true,
+          // ignore: avoid_redundant_argument_values
+          includeSettingsArtifactHub: true,
+          // ignore: avoid_redundant_argument_values
+          includeSettingsDocumentationArtifactHub: true);
       writeFakeManifest(example, isApp: true);
 
       final List<String> output =
@@ -787,6 +865,71 @@ dependencies {
       expect(
         output,
         isNot(contains(GradleCheckCommand.exampleRootGradleArtifactHubString)),
+      );
+    });
+
+    test('prints error for declarative method of applying gradle plugins',
+        () async {
+      const String packageName = 'a_package';
+      final RepositoryPackage package =
+          createFakePackage('a_package', packagesDir);
+      writeFakePluginBuildGradle(package, includeLanguageVersion: true);
+      writeFakeManifest(package);
+      final RepositoryPackage example = package.getExamples().first;
+      writeFakeExampleBuildGradleGP(example,
+          pluginName: packageName,
+          includeBuildArtifactHub: true,
+          includeSettingsArtifactHub: false,
+          includeSettingsDocumentationArtifactHub: true);
+      writeFakeManifest(example, isApp: true);
+
+      Error? commandError;
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['gradle-check'], errorHandler: (Error e) {
+        commandError = e;
+      });
+
+      expect(commandError, isA<ToolExit>());
+      expect(
+        output,
+        containsAllInOrder(<Matcher>[
+          contains(GradleCheckCommand.exampleSettingsArtifactHubStringGP),
+        ]),
+      );
+      expect(
+        output,
+        isNot(
+            contains(GradleCheckCommand.exampleRootSettingsArtifactHubString)),
+      );
+    });
+
+    test('error message is printed when documentation link is missing',
+        () async {
+      const String packageName = 'a_package';
+      final RepositoryPackage package =
+          createFakePackage('a_package', packagesDir);
+      writeFakePluginBuildGradle(package, includeLanguageVersion: true);
+      writeFakeManifest(package);
+      final RepositoryPackage example = package.getExamples().first;
+      writeFakeExampleBuildGradleGP(example,
+          pluginName: packageName,
+          includeBuildArtifactHub: true,
+          includeSettingsArtifactHub: true,
+          includeSettingsDocumentationArtifactHub: false);
+      writeFakeManifest(example, isApp: true);
+
+      Error? commandError;
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['gradle-check'], errorHandler: (Error e) {
+        commandError = e;
+      });
+
+      expect(commandError, isA<ToolExit>());
+      expect(
+        output,
+        containsAllInOrder(<Matcher>[
+          contains(GradleCheckCommand.artifactHubDocumentationString),
+        ]),
       );
     });
   });
